+++ src/Products/PloneMeeting/content/organization.py
@@ -14,7 +14,6 @@ from imio.helpers.content import get_back_relations
 from imio.helpers.content import uuidsToObjects
 from plone.autoform import directives as form
 from plone.dexterity.schema import DexteritySchemaPolicy
-from plone.memoize import ram
 from plone.supermodel import model
 from Products.PloneMeeting.config import PMMessageFactory as _
 from Products.PloneMeeting.events import _invalidateOrgRelatedCachedVocabularies
@@ -184,7 +183,7 @@ class PMOrganization(Organization):
 
     def get_acronym(self):
         """Accessor so it can be called in a TAL expression."""
-        return self.acronym
+        return self.acronym or u''
 
     def get_groups_in_charge(self, the_objects=False):
         """Accessor so it can be called in a TAL expression."""